got dashboard working. started work on admin dashboard and password hashing
