+++ Software_Engineering_Project/Views/Shared/AdminDashboard.cshtml
@@ -1,0 +1,24 @@
+﻿@model Software_Engineering_Project.Models.Users
+<div>
+    <h2>Admin Dashboard</h2>
+    <h4>Hello @Model.name</h4>
+
+    <br />
+
+    <h5>Users</h5>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+
+    <br/>
+
+    <h5>Rooms</h5>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+
+</div>
