--- Software_Engineering_Project/Views/Shared/Dashboard.cshtml
@@ -1,4 +1,0 @@
-﻿@model Software_Engineering_Project.Models.Users
-<div>
-    <h3>Logged in as @Model.name</h3>
-</div>
