+++ Software_Engineering_Project/Views/Shared/UserDashboard.cshtml
@@ -1,0 +1,24 @@
+﻿@model Software_Engineering_Project.Models.Users
+<div>
+    <h2>Dashboard</h2>
+    <h4>Hello @Model.name</h4>
+
+    <br />
+
+    <h5>Current Bookings</h5>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+
+    <br/>
+
+    <h5>Previous Bookings</h5>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+    <p>blah</p>
+
+</div>
