+++ Software_Engineering_Project/Views/Shared/_Layout.cshtml
@@ -22,17 +22,20 @@
                 @Html.ActionLink("Booking System", "Index", "Home", new { area = "" }, new { @class = "navbar-brand" })
 
             </div>
+            <!--
             <div class="navbar-collapse collapse">                
                 <ul class="nav navbar-nav navbar-right">
                     <li class="dropdown">
                         <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false"> Login <span class="caret"></span></a>
                         <div class="dropdown-menu" style="padding: 15px; ">
-                            @Html.Partial("dropDownLogin", new Software_Engineering_Project.Models.Users())
-                            
+                            @{ 
+                                //Html.Partial("dropDownLogin", new Software_Engineering_Project.Models.Users())
+                            }
+
                         </div>
                     </li>
                 </ul>
-            </div>
+            </div>-->
         </div>
     </div>
